Adapt make-disk-image to our requirements
- use GPTBIOS boot layout
- 'root' is partition 1
- hard-code XFS since we won't use ext4

It was a bit tricky to get an XFS filesystem into a partition starting
at 4MiB. In contrast to mkfs.ext4, mkfs.xfs does not support an --offset
flag. Furthermore, the sandbox does not allow to set up loopback
devices which would be an easy option to access an image starting at a
fixed offset. To work around these limitations, I create an XFS
filesystem in a separate image first and copy this image into the
partition of the final image. Thanks to direct I/O reads and sparse
writes this is less expensive than one thinks.