Avoid logrotate errors on fc-manage runs
* try to restart nginx after log rotation if sending the signal fails
* remove logrotate from multi-user.target which avoids running logrotate
during fc-manage runs

Logrotate produced an error during a fc-manage run which caused the
maintenance task to fail. This happened because nginx was restarted at
the same time and logrotate tried to send a signal to the missing process.
Each one of the changes fixes the problem but it makes sense to include
both to avoid errors in similar situations.

bugs id: #119218