Use our pinned nixpkgs version when building releases on Hydra
It's ok to use the current revision of nixpkgs via <nixpkgs> for bootstrapping,
but we don't want to build with that version.
The release function has a nixpkgs arg which is by default our pinned
version, but it is overridden by Hydra because it has the same name.
Renaming fixes the problem.

bugs id: #122759