Fix statshost metric relabeling config
We used the wrong prometheus config attribute relabel_configs which is
run before scraping a target. We want metric_relabel_configs. That
option was missing from the service code from upstream, added it.
Also, in some places drop was used, but labeldrop was meant.
This adds action options that are missing upstream which includes
labeldrop.

bugs id: #122902