Fix rabbitmq timer execution and warning
The service had RemainAfterExit=true set which means that it stays
active all the time. Timers don't reactivate active units, so it only
ran once and not hourly as intended.

The executed unit must be specified with extension, for example .service.
Systemd complained about that (no valid unit type) and ignored the line.
The default service name is the same as the timer name, so we don't have
to specify it at all.

bugs id: #123236
bugs id: #123185