MySQL: set root password from password file on startup
We want to be able to change the mysql root password without writing it
to the Nix store. The password from the password file (default location
/etc/local/mysql/mysql.passwd) is set from the pre-start script now.
It also creates /root/.my.cnf. This ensures that the files containing
the password always match the actual password (after a restart).
If the password file is missing on startup, it's generated randomly.

bugs id: #123366