Add support for multi-node Kubernetes (experimental)
Uses Kubernetes 1.15 from NixOS 19.09.

Roles:

* kubernetes-master
  * no pods, exactly one per RG
* kubernetes-node
  * joins existing cluster in RG automatically
  * runs pods, at least one per RG

kubernetes-master and kubernetes-node can be combined on a host.
Activating kubernetes-node on an existing master-only VM works, but taints must
be disabled manually to run pods on it.

Policy routing is disabled on Kubernetes VMs because it interferes with
routing of Kubernetes service IPs.

Other VMs in a RG with a Kubernetes cluster can also access service IPs
via a route set to the first Kubernetes node.

Dashboard and API allow public access for authenticated users.

sudo-srv users can generate a kubeconfig usable for kubectl and the
dashboard by running `kubernetes-make-kubeconfig` on the master.

Case 122936