Reset maintenance flag every time
Every time fc-agent is called with '--maintenance', the "in maintenance"
directory flag gets reset at the end of the run unconditionally.

The maintenance client used to reset the "in maintenance" directory flag
only if a maintenance job actually has been run. This meant that in case
a maintenance job aborts, the flag does not get reset until another
maintenance job is created. This may take a long time on VMs not running
updates in maintenance.

Also increase maintenance archive retention time to assist debugging of
old cases.

Case 119218