Create a new role for LAMP application servers.
Security:

We open ports 8000 freely but we don't expose them on ethfe, only on
localhost and ethsrv (relying on the firewall). The apache user is
limited to 'nobody:service' and access to docroot and upload directories
 et.al. is granted through the application deployment.

Tideways is bound to localhost so PHP can talk to it, it talks to the
internet in turn, when activated by the application deployment.

Tested by manual inspection.

Create a new role for LAMP application servers.

Security:

We open ports 8000 freely but we don't expose them on ethfe, only on
localhost and ethsrv (relying on the firewall). The apache user is
limited to 'nobody:service' and access to docroot and upload directories
 et.al. is granted through the application deployment.

Tideways is bound to localhost so PHP can talk to it, it talks to the
internet in turn, when activated by the application deployment.

Tested by manual inspection.