mailserver: dynamicMapFiles
The new option flyingcircus.roles.mailserver.dynamicMapFiles consists of
a set of map which will be dynamically compiled on `systemctl restart
postfix` and do not need a full fc-manage run. The attrset's key is the
postconf parameter and the value is a set of files in postmap format.
The mentioned files will be automatcailly made writable by service
users.

Note that `systemctl reload postfix` does currently not cause the files
to be compiled. This will be handled in a separate follow-up case 127993,

Also introduce /etc/local/mail/config.json.

Basic configuration including list of domains, mailHost and
webmailHost gets now included from `/etc/local/mail/config.json`. It
used to be set from a custom nix file in /etc/local/nixos.

Still need to update role docs. This will be solved as part of
case 127685.

Case 126884