Include part of github rev in Nix store name
This is a safety measure to avoid accidentally missed version updates.
Fixed output derivations are referenced with their derivation name and
their sha256. We used to have fixed derivation names like "nixpkgs" so
that the Nix store path was depending only on the sha256 sum. If a
developer updates the rev but forgot to update the sha256 sum, the
deriation was not changed at all.

We now append part of the rev to the derivation name so that this kind
error will be caught early on.