Increase max. workers for Apache.
With actual customer peak load we may want to let more requests through.
This number is thought of as a ceiling what a single VM (maybe
with multiple cores) could handle.

If you have smaller VMs or can't handle that much load then the
load balancer should be limiting the number of requests given to a
single Apache. That's why the spare servers are set relatively low in
comparison with the MaxRequestWorkers.