Sensu check for nginx worker age: add a non-critical period
Before, the check went critical for every worker that was older
than the config file. This is too noisy. Nginx may take
some time to switch the workers to shutting-down mode when the system is
overloaded or swapping, for example. Another reasons for outdated workers
may be a config error that was not caught by the check script or an
incompatible listen directive change.

Case 127883