Remove `/result` symlink after 3d
After discussion we came to the conclusion that / is never a good place
for a result symlink and such links should be automatically removed
after a grace period. `result` links in /root or any other home
directory are not touched.

We use check_age to dermine if the link can be deleted. tmpfiles.d and
other mechanisms are not able to measure the age of the link (and not
its target).

Case 128533