Agent: use spread time in channel-with-maintenance mode, check every hour
Spread time was already in use for the channel update mode which we use
on non-prod VMs. Production VMs, however, first build the new channel
and schedule the switch in a maintenance window. Spread wasn't used in
that case. That meant that all production VMs built the new channel directly
after the release in a short time window. Building is quite
resource-intensive and can cause overload on our storage systems.
Building the new channel now also uses the spread time.
It's now 60min instead of 120min to roll out channel updates faster.

Case 127919