Kubernetes: run dashboard and coredns as systemd services on the master
On NixOS, dashboard and coredns run in the Kubernetes cluster while all
other Kubernetes core services run directly on the master as systemd
services. We had various problems with coredns and dashboard that were
hard to debug, mostly related to secrets management.
The new setup is less complex and should be more reliable.

The master node runs coredns which acts as nameserver for pods in the
cluster and other VMs in the Kubernetes RG.

This also adds some assertions that a master is present in the
RG.

Dashboard now does HTTP Basic auth using the generated
htpasswd_fcio_users file.

Case 128939