Temporarily exclude 'tested' from channel prereqs
To get up VMs fast, we publish channel updates without requiring the
'tested' target to succeed. Tests are executed anyway.

The aim is to get VMs up and running now and fixing remaining tests
later on. Of course this has to be done before we use 20.09
productively.

Also fix garbagecollect test and images.fc.