Hotfix: allow Nginx master process to read all files
Commit d8f04ff87d68b15057dc3979dd460672d73f5cb4 meant to give
root permissions to the master process to allow access to protected
files like certificate keys by starting it as root but that isn't
enough. The capabilities set in the SystemD service prohibited
access to files that don't have permissions that would allow the root
user to access it. Setting CAP_DAC_READ_SEARCH allows the process
to read these files.

PL-129520