Nginx: reuseport for structured vhost config, also recommended for plain config
listen ... reuseport ... starts a separate socket listener for each
Nginx worker process instead of having only one in the master process.
This is meant for higher performance but also allows worker reloading
when the listen IP changes from 127.0.0.1 to 0.0.0.0, for example.

Add an `reuseport` option for `listen` which is on by default.
This affects structured JSON and Nix config.

For plain config files, the option is recommended in the documentation and
example config.

 #PL-129589