Kubernetes: add FC nameservers to resolv.conf as fallback
resolv.conf on Kubernetes VMs only used the CoreDNS on the master as
nameserver. CoreDNS forwards requests to FC nameservers.

Also adding the FC nameservers to resolv.conf avoids annoying failures,
for example when the acme client wants to contact
the letsencrypt servers on a fc-manage run when CoreDNS is restarted at
the same time.

 #PL-129590