Nginx: ensure correct logfile permissions, logrotate fixes
We always want the logfiles to be writable by root or restarts
will fail. The master process runs as root but due to restrictions
enforced by SystemD it can only open its own files for writing.
Nginx creates and/or chowns logfiles when:

* reopening log files after logrotate (SIGUSR1)
* checking the config
* startup
* reload

The results vary. Some variations don't break things
but we force root:nginx everywhere now for consistency.

The automated test covers multiple situations I could think of.
Let's hope that this is enough for now...
Not running Nginx as root should make things easier but that
would need some more work.

Logrotate now uses the correct permissions and uses reload instead of
restart if log file reopening failed to avoid restarts that may fail and
leave us without a running Nginx.

The maximum age for log files now uses the rotateLogs setting so
tmpfiles doesn't delete log files that should still be kept around.
The fixed value of 10 meant that rotateLogs > 10 (days) had no
effect.

Also adds logrotate to the system packages and adds a command
logrotate-config-file to show the location of the config file which
makes debugging logrotate config easier and is needed for the test.

 #PL-129618