Journalbeat: tune timeout/ttl to avoid connection errors
Journalbeat keeps the TCP connection open and may send nothing
for a long time when there's no new message in the journal.
HAproxy closes the connections after the timeout and Journalbeat
complains about the closed connection.

This increases the timeout on the HAproxy side to 121s and
lets Journalbeat reconnect every 120s to avoid being thrown out.

 #PL-128336