HAProxy: fix 19.03 upgrade problems
Disable the stats socket line from the 19.03 example config
if it's still present in local config. It breaks the service now.

We need to trigger a restart if the haproxy socket at the new
location is missing. This is the case when upgrading from 19.03
or when something accidentally deleted the socket.

Test ported to Python and extended to check reloading.

 #PL-129632