Agent: only switch to new system if config changes (lazy activation)
fc-manage now has a -l flag that skips the switch if the build
result is the same as the running system.

The flag is used by default for automatic agent runs and can be
disabled with the NixOS option flyingcircus.agent.lazySwitch.

This saves some CPU time and avoids triggering targets on every agent
run which is a problem for the ACME renewal targets, at least.

May have some side effects because tmpfile rules are not executed every
10 minutes and inactive units that should run will not be activated
anymore. If something like that is really needed, an explicit timer
or unit settings like Restart=always should be used to trigger that.

Also adds an option flyingcircus.agent.verbose to enable debugging
output.

 #PL-129745