Servicecheck: use a timer to start the unit, monitor updates
Wasn't started at all because it depended on fc-manage.service which is
called fc-agent since 19.03. Using a timer is better because we will
switch to an event-based agent run model later.

A sensu check now checks the update time of the servicecheck file
which is expected to be updated (or at least touched) every ten minutes.

 #PL-129738