Coturn: fix ACME cert renewal
Using the group option of the ACME certificate doesn't work as expected on
20.09 anymore. Previously, setting it to turnserver enabled coturn to see
the cert. Now, the group is used as the group for the ACME challenge file
which must be readable by Nginx. Changing it made the file inacessible
for Nginx. This is fixed by using ACL instead to allow the turnserver user
to see the certificates.

 #PL-129729