fc-agent: prevent garbage collection of system we are switching to
If the switch takes a bit longer than normal, there's a realistic chance
that the garbage collector removes the system we are just switching to.
This can cause serious problems for the system. We must keep a link to
the system to have a GC root until the system is registered in the
profile.
The link will be kept at /run/fc-agent-built-system until the new system
is registered or until the next reboot if registering failed.

 #PL-129765