fc-agent: fix pre-building for maintenance, improve update info
nixos-rebuild dry-build actually built the current system, not the next
channel as expected so the building happened later in the maintenance activity.
Also, the list of changing services was missing from the
maintenance activity message because of that.

The error exists since 19.03. Back then, the error message that
default.nix is missing in the combined channel
(release channel at /root/.nix-defexpr/channels/next) was discarded. It
"worked" without an error on 20.09 because there's a default.nix in the
combined channel but it still isn't used by nixos-rebuild because the
directory structure is wrong. It's just ignored and nixos-rebuild goes
to the next NIX_PATH entry
/nix/var/nix/profiles/per-user/root/channels/nixos which is the current
channel.

The combined channel is a directory that contains nixpkgs but isn't
nixpkgs so -I nixpkgs=(...)/next should be -I (...)/next. Nixos-rebuild
now correctly finds the new nixpkgs inside the dir and builds it.

The message now also includes the new environment name and the Hydra
URL.

The string representation of Channel is now more helpful
and consistent even with local checkouts.

 #PL-120275