Remove docker from kubernetes test and use containerd
Kubernetes dropped docker support, so our test must be changed to remove
all commands relaying on docker. Containerd is used by kubernetes
instead and must be used in the test to not break it.

This also reactivates the kubernetes test because it is not
broken anymore.

Also deaktivates a subtest, because it currently does not work on 21.05
but it is not essential.

   # PL-129854