Add shell.nix which as dev-setup alternative
Sets up NIX_PATH to include the current fc-nixos dir and the
pinned upstream channel sources. Can be used with nix-shell
in the simple case or direnv/lorri for a more comfortable dev
experience. Doesn't help with fc-manage yet because it still expects
a channel dir which cannot be created from a shell.nix (sandbox).