Revert "Adjust filename for ClamAV's daily signature file at our sensu check"
This reverts commit ff44ef6e85b4262168f31993b11a277cdc91b0b2.

This fixed a temporary problem that went away with the next signature
update. daily.cld wasn't present for a day, but daily.cvd was.
In normal operation, we expect daily.cld to be there and thus we should
monitor that.

We have to improve the check later to handle that exceptional case.