ceph/mon: restructure config, fix minor issues
We now do not reload/restart upon changes to the config or fc-ceph.
I noticed way too many restarts for production use and we might be
able to do this in the future with a newer Ceph version but for
now we are used to this gap anyway.

Improve logging output for timers: Python normally buffers its output
and then we can not see anything while a timer job is progressing.