fc-agent: show significant line of nix-build error in log msg
Nix build errors appear at the end, with a potentially long stack trace
before it. The interesting bit of the error message may not be last
because there are some errors that display additional info. In Graylog,
one needs to drill down in the log message to find it, the message
itself is just a generic "build failed".

Find the most interesting line and use it directly as the log message
so we see the error in the overview. Also helps when viewing logs with journalctl
with standard settings.

 #PL-130088