fc-agent, garbage-collect: limit IO and CPU usage via unit settings
We experienced some load alerts for short peaks caused by garbage
collection or the agent. This is often caused by iowait and it seems like
that VMs that are mostly idle are affected more than others.

Restrict the number of IOPS on cgroup level depending on the limit for the
VM to avoid requesting much more IOPS for maintenance tasks than the VM can deliver.
The agent gets more because it shouldn't take forever. For the garbage
collection, there's a more restrictive setting because time doesn't
matter here and the timeout is set to infinity now.

We also set other IO and CPU-related unit settings to make sure that regular user
processes have higher priority than our maintenance.

Drop the limitation on how much fc-collect-garbage can collect in one
run because the other limits should prevent problems.

 #PL-130083