raid: disable check for other error count
The check triggers at "1 other error count" which is way too tight.
We have enough redundancy concepts to wait for actual problems and
Thomas Krenn used to say "we change disks at 50 errors and up".

The script is too hard to refactor for the 50 errors rule, so we
just use the existing toggle to disable the "other error" count.

Fixes PL-130057