fc-agent: use structlog in reqmanager
Used stdlib logging before so we lost the log messages when reqmanager was called
by fc-manage. This is the case for our automatic agent runs which means
that we didn't have log messages with the id of failed maintenance requests, for example.
Using structlog in reqmanager fixes this and also gives a lot more insight into what's
happening when maintenance requests are handled now.

 #PL-130088