Use OpenSSH 8.7 for ssh/sshd
Use the newest openssh version to fix frequent sshd subprocess crashes
which are caused by failing malicious connection attempts.

sshd tries to log something after a timeout but is killed
because syscall 31 is blocked by seccomp.

This doesn't override the openssh package globally because this
would cause a big rebuild. We just use it for ssh and sshd by
setting programs.ssh.package.

 #PL-130095