Make option to change warning and critical threshhold for sensu disk usage check
In some cases our set percentage of 95% and 90% for disk usage warning
and critical may be to high or to low.
An application may internally implement a diskusage check at lower than
90% and abort an action if that threshhold is passed. In that situation,
our monitoring does not warn us fast enought.
Otherwise a disk may have such high capacity that even at a usage
percentag of 90% there is nothing to worry about for quite a long time
and a warning then would be just noise.

Relates to FogBugz ticket 143476.