Sensu server: fix breaking rabbitmq update script
The check for existing users (name is the fqdn of the node)
was only doing a partial match which could match the wrong node fqdn,
thinking that the client already exists. Skipping the user creation
breaks the script in the next step. Matching from the beginning of the line
fixes it.

This affected even more new nodes that are sorted after the breaking node.

 #PL-130123