Nginx: set ambient capabilities when masterUser is not root
Nginx behaves differently when the master user is not root.
In this case, we use the ambient capabilities from upstream to
allow binding to non-privileged ports.

This is a follow-up to c98d4dd09caf1a8049320e3bd84636a15cf3a7e3
and doesn't affect our nginx module which uses root anyway.