Build PRs in Hydra
This commit creates a couple of .hydra configuration files.

The project.json defines a jobset called ".jobset", which will run
a program (hydra-github-jobsets-generator) to generate all the
other jobsets.

That program loads the list of pull requests provided by the
githubpulls input, and the `.hydra/config.json` file from the
generator_config input to produce a jobset per PR.