k3s: install separate kubectl binary to speed up CLI interaction
Currently, kubectl is aliased to k3s kubectl which takes some time to
start up because it bundles all kubernetes components. Installing a
separate kubectl binary makes interaction with the CLI much faster.
Still works the same because kubectl interprets KUBECONFIG which is set
to the kubeconfig that k3s kubectl uses.

 #PL-130240