Update backy-extract to 1.1.0
This version has optimized locking and memory usage.

backy, backy-extract and restore-single-files are now included directly in the
backyserver role, instead of the overlay because importing the Nix expression
from the backy-extract repo breaks nix-env -qa (path is not valid) in
our channel test. These tools make only sense if backy is installed on
the system anyway.

 #PL-130044
 #PL-129786