sensuserver: increase kernel neighbor table size/gc limits
We hit the default size limit of the table that caches ARP responses on
our rzob sensu machine because it talks to a lot of machines now.
This may have been the cause for sensu API timeouts and produces
error messages in dmesg on the sensu server:

neighbour: ndisc_cache: neighbor table overflow!

Increasing the limits also avoids unneccessary garbage collection.

 #PL-130235