matomo: don't require a local database
The setup and archive processing units had *requires* relationships to
mysql which may or may not be present on the local system. Matomo can
be configured to use a database on another host which causes the
services to fail if mysql is not present locally. Relax the dependency
and use *wants* instead so mysql is started if it's present, otherwise
ignored.

 #PL-130234