Update nixpkgs, adapt journal shipping after journalbeat removal
journalbeat is no longer available in ELK 7.16 and must be replaced
with filebeat with the journald input. This change keeps the
flyingcircus.journalbeat option but implements it with filebeat now.
Because filebeat uses a different way of saving the current cursor
position, this includes a script that migrates the old format to
the new format and can create the new format from scratch on service
start.