filebeat: finish migration for journalbeat registry file
Add the old journalbeat state directory as second state dir for the
filebeat-journal-* service to make it easier to fetch the old registry
file. By using DynamicUser, accessing state dirs of other services is
not possible (at least there's no easy way).
The old state dir stays as an empty dir, we can remove it when all VMs
are upgraded.