gitlab: fix assets for non-standard gitlab packages
Use the configured package from the gitlab service for assets delivered
by Nginx.

We always used assets for pkgs.gitlab, even if the package
has been changed for the gitlab service. This fails completely
for Gitlab EE because assets have different hashes and no CSS and JS
can be loaded anymore, for example.