statshost: InfluxDB optional, option for scrape interval
InfluxDB will go away with 22.11, for now we need options to remove it
gradually.
There are 3 new options to completely disable InfluxDB or just disable
reading/writing from/to influxdb.

Prometheus gained an option to change the prometheus scrape interval.
This can be used to add a second (or more) statshosts in an RG with
reduced sample rate but longer retention time as InfluxDB replacement.

 #PL-131027