libmodsecurity: 3.0.4 -> 3.0.6
Vendored libmodsecurity package from nixos upstream with disabled
lmdb support. With it, nginx segfaults in NixOS tests and database
files are dropped in the current directory where you reload nginx.
We didn't use lmdb in the previous version and libmodsecurity
builds without it by default, so we should be fine without it.

 #PL-130446