Disable createHome for users
This avoids setting the permissions for home dirs twice in the
activation script. This led to situations in the past where the
activation script died after the users script and home dir permissions
were set to 700.
Let our activation snippet handle home dir creation and
setting permissions to 755.

Add a test to check home dir creation.

 #PL-130524