fc-collect-garbage: ignore errors in systemd unit
It's a common problem that fc-userscan dies when files are deleted while
it is running. This failure now marks the fc-collect-garbage unit as failed
which may break a system rebuild. The reason is that the
switch-to-configuration script is looking for units that failed during
the system switch and that it aborts if it finds one.
This doesn't mean that the failure has been caused by the switch, it may
be just a coincidence. We have seen this problem when automated deployments
delete files and also start a system rebuild while the
fc-collect-garbage unit is running concurrently, activated by its timer.

The unit now accepts exit codes 0, 1, 2, 3 as "successful" exit codes
so it doesn't trigger the problem anymore. We can tolerate errors in the
garbage collect unit because it will be run on the next day again.

 #PL-130101