Elastic: use OSS versions of 6.x, add OSS versions for 7.x
Elasticsearch/Kibana6: OSS (Apache 2.0) still present on 21.11 as
elasticsearch6-oss and kibana6-oss, use them by default in the role code.

Elasticsearch/Kibana7: pin to 7.10.2 which is the latest version
available as OSS, use these packages by default in the role code.
Version 7.10.2 is the same as on our 21.05 platform and it's the
newest available OSS version of 7.x.

 #PL-130528