matomo: fix potential Nginx path traversal, home dir permissions
The state dir is also the home dir for the matomo user and has
createUser set. This setting changes permissions from the
setup script when system activation runs and the
wrong permissions remain when the setup script does not run after that.
createHome can be disabled, the setup script takes care of everything.

The /js/tagmanager location was missing the trailing slash which makes
path traversal possible.

 #PL-130561