Selectively allow unfree packages
Do not enable allowUnfree anymore so we don't miss critical license
changes in the future and have to be explicit about what we want to
allow in platform code or for specific machines/projects/customers.

Add an option flyingcircus.allowUnfreePackageNames which can be extended
from multiple locations.

megacli is needed for the platform. We also allow mongodb for our
customers as its SSPL license is rather unproblematic as long our
platform code is SSPL-compatible.

Hydra is still allowed to build all packages.

 #PL-130528