Agent: fix bootstrapping for new VMs
Because of an inverted file existence check, initial ENC data was never
loaded on bootstrap, rendering new VMs unusable. The failure to load ENC
data also uncovered a problem with the fallback channel URL which is
used when there's no channel URL in the ENC data. Redirects for that
URL are resolved now before passing it to nix-build which doesn't work
with redirects.

 #PL-130631