Remove broken check for obsolete result links in /root
Generates monitoring noise and moves statuspages to "unknown" state so
we should get rid of it. sensu-client is not allowed to check the link
in /root anymore on 21.11 and we don't want to have security expections
just for that. We need to find another way for checking obsolete result
links as they can be everywhere and /root is just one place.

 #PL-130493