antivirus: quick and reliable database initialization
With the previous PR, the clamav daemon still may crash on the first
run because the database is not preset yet.

Setting OnActiveSec to a low value or even 0 for the freshclam timer
doesn't help when randomization is active which which delays the
start for up to 60 minutes.

The new solution here introduces a new unit which is only triggered
when the required database files are absent. The opposite condition
prevents the daemon from starting and crashing instantly, until the
database fetch is finished.

The NixOS test now asserts basic cases for the database sensu check
and makes sure that the private mirror is set up correctly.

 #PL-130648