fc-ceph: clarify naming of sub parser bindings
During the creation of subparsers and their commands, the name bindings
of individual subcommands have been re-used. While that is okay in
principle, the actual naming of these bindings was confusing in a
particular case.
Additionally, a comment explaining that this re-binding of names is okay
has been added.

This is a merely cosmetic code change and should not alter any
functionality of the tool.