The xfs bug check is too coarse and results in floods of false positives
When Ceph had a global IO block we saw this check (and this will now
be red until the next reboot) on a lot of VMs that do not exhibit the
known XFS bug (PL-130162).

Use a more specific marker to grep for and provide more context
in the check's output.

Port to 21.11 and 22.05.

(See PR #557)