tests/ceph: improve output in failure situations
We noticed that failures during fc-ceph mon create went unnoticed
(due to using `execute` instead of `succeed`) and also that no real
output was provided. It appears that the &> redirection pattern isn't
properly supported here so we replaced it with the more typical 2>&1
pattern.

Also, avoid showing the journal output after creating the first mon
too early as it initially shows the (expected) failures of the
non-bootstrapped mon and that can be quite confusing.