fc-ceph: refactor how we call external commands
We noticed that black continuously formats external command calls
in a basically unreadable fashion and started introducing fmt:off
blocks.

At the same time we had a multitude of different run helpers
and even residual calls not leveraging JSON interfaces from back when
this was running on Gentoo.

I created a runner structure that simplifies the Python code on the
application side and also provides a structured way to interact with
JSON CLIs.