Add fc-postgresql command and autoupgrade options
The new command `fc-postgresql` (to be run as postgres user) can show
the current state of data directories for the available major versions,
prepare and run upgrade migrations.

Documentation now shows how to prepare and run manual upgrades using
`fc-postgresql upgrade`.

This also adds the `flyingcircus.services.postgresql.autoUpgrade`
submodule which is at the moment intended to be used internally in
platform code for roles that need postgresql and want to manage upgrades
without user intervention. The options are not in the user documentation
but we may add them later when we want to recommend using auto-upgrades
to users.

Documented fc-postgresql commands are:

- list-versions
- upgrade

Undocumented in customer docs:

- check-autoupgrade-unexpected-dbs (to be used as Sensu check if
  autoupgrade is enabled)
- prepare-autoupgrade (to be used manually to prepare automated
  upgrades)

 #PL-130197