java: use crc32-patched zlib for more openjdk variants
We only used the zlib patch for some of the headless versions before and
missed the aliases, so for example jdk11_headless used the patch while
openjdk11_headless war not despite being they the same package upstream.
We now have a test that makes sure that we don't break the aliases and
patch all openjdk versions and their headless variants.

Java 17 wouldn't build with our patched zlib which was due to a missing
zlib patch from upstream. Our crc32 patch is now added to the list
instead of replacing it.

 #PL-130760