tests: remove subtest which build the system
It's too hard to maintain the code that is needed
to sync the situation in the VM with the system built
on the outside. It now tries to build a kernel but there's
no obvious reason why this happens now.

We now have a series of smoke test VMs that can check that
building systems with local config works properly.