jitsi: add watchdogs for jicofo and jvb, use notify service type
Jicofo and jitsi-videobridge2 services now have "watchdogs" in their
start scripts that restart the service when timing out after 40 seconds
by using WatchdogSignal=SIGTERM and Restart=Always.

Both use the Jicofo health API but they check for different things: The
videobridge watchdog checks if the videobridge is unusable for Jicofo
while Jicofo itself checks for everything else (like 5xx codes from the
API or connection timeout).

Both services now use the "notify" service type, signalling their
readyness only when the Jicofo API health check is ok (from their
perspective).

Because of that, the separate service for delaying the videobridge
start should not be needed anymore and is removed.

The videobridge also uses "Wants" for the relation to Jicofo now
so it doesn't die immediately when Jicofo exits but still starts
Jicofo if it's not active when videobridge wants to start.

Other changes:

* don't stop Jicofo when the unit changes
* fix shellcheck warnings for unquoted vars and `export var=$(...)`

 #PL-130720