Fix fc image building
Needed a newer lkl/linux version which is compatible with current
xfsprogs and more RAM for the image building VM.

Allow cptofs to use more RAM by patching the C code that starts
the kernel.

cptofs limits the memory usable by the kernel to 100M which
seems to be enough when I run the image build manually but not
when run by Hydra. Doesn't really make sense but allowing
cptofs to use 500M of the available RAM won't hurt anyway.