postgresql: remove obsolete fix for postgis problem, cleanup
Postgis seems to work properly in the test now for 12, so we can
simplify the test script.

Replace rolename with a version string. This is possible now because
the last version with the old naming scheme, 9.6, got dropped and
versions start at 10 now. We use a version argument most of the time,
it makes version comparisons easier and is less to type.