Remove some global pkgs or replace with non-X variants
This reduces system closure size by about 500MB for basic VMs.

Adding packages to environment.systemPackages is easy today with
/etc/local/nixos config and many of the tools can be also ran via
nix-shell if needed. App dependencies should be part of app deployments
and not be expected to be present globally.

* python2Full is big (pulls in a lot of graphics stuff) and should not
  be used anymore.
* Installing libraries globally has no effect anymore.
* links2 and latencytop now don't depend on graphics libraries anymore.

Some X libraries are still in the base system because they cannot
be easily removed from everything. There's an `environment.noXlibs`
option but it has too many side-effects and needs a lot of custom
compilation on our Hydra.