shell.nix: fix NIX_PATH in shell, use sudo in nixos-repl
Just passing a env var to mkDerivation is not enough to change the
NIX_PATH inside the shell when run by nix-shell. We have to
export it from the shellHook.

Before, this worked only for nix-direnv.

We have to run nixos-repl.nix as root, adding sudo to the nixos-repl
command is more convenient. Using `-E` keeps the environment to
use the correct NIX_PATH set by the shell.

(cherry picked from fc-22.05-dev commit 417060efaa45d945e9486fabf57199c92bff6021)