webgateway: update Nginx proxy timeouts and HTTP version
Use HTTP/1.1 to communicate with upstream servers. This fixes HTTP 413
responses from haproxy when Nginx passes a GET request with a
request body. haproxy stopped to accept these requests via HTTP/1.0
starting with version 2.5 because they could be a security risk.

Add the `proxyTimeout` option from upstream and use its default of 60s
instead of 90s which is also more in line with the recommendations from
the Nginx project.

Also adds some uses of literalExpression from the upstream module for
option examples and default texts.

 #PL-130875